pyproject.toml
[project]
name = "python-intermedio-venv-pep8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
]
